server updates and fixes.
