+++ server/package.json
@@ -7,7 +7,7 @@
     },
     "dependencies": {
         "dotenv": "^16.3.1",
-        "nanoid": "^3.1.20",
+        "nanoid": "^3.3.6",
         "node-cache": "^5.1.2",
         "socket.io": "^4.7.1"
     },
@@ -17,8 +17,8 @@
         "start": "node ./lib"
     },
     "devDependencies": {
-        "@types/node": "^14.14.21",
+        "@types/node": "^14.18.54",
         "ts-node": "^9.1.1",
-        "typescript": "^4.1.3"
+        "typescript": "^4.9.5"
     }
 }
